Fix flake in CallAudioRouteControllerTest#testToggleMute
It's possible in some test runs for the last verification of onCallAudioStateChanged to be invoked more than once. Looking at the failed logs, there are only 3 invocations so it's possible that the first invocation is being processed as part of the 3rd invocation when we verify this the last time as well. Changing to verify atLeastOnce() will resolve this.

Bug: 395098656
Change-Id: Id11ad9db882351c238d9baf7d9704817dec27719
Flag: EXEMPT test fix
Test: atest CallAudioRouteControllerTest