Do not add SPEAKER_ON message if the route has been on the speaker
If the audio route has been already on the speaker, the SPEAKER_ON
message should not be added. This typically occurs when the active
state changes. The fix allows the route to exit faster, and avoids
an unexpected message when a speaker off event occurs during the
pending state.

Flag: EXEMPT bugfix
Bug:392012245
Test: manual
Change-Id: I3f84e0f2588eda0f5825c4382749763c967d3009