Dsda: Resolve answering MT drops unholdable FG call
Ensure that when we have an unholdable fg call that we add the
Connection#EXTRA_ANSWERING_DROPS_FG_CALL to the incoming call connection
extras. This will ensure that Dialer will have this available in order
to show to the user in the call screen that answering the call will drop
the currently foregrounded call.

Bug: 394900917
Test: atest CtsTelecomCujTestCases:CallSequencingBasicCallTest
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: I05fbbfa692ba6cb314c8ac0f64351bd25f024f21