DSDA: Disconnect nonholdable calls with MT call
Ensures that we disconnect the non-holdable calls on one sim when we
receive an incoming call another sim. Previously, we would just
disconnect the active call but in a scenario where we have a HOLDING +
ACTIVE call plus a RINGING call (on another sim), we should also opt to
disconnect the held call as well. A good example of this is with
Verizon.

Bug: 392698669
Test: atest
CtsTelecomCujTestCases: CallSequencingManagedHoldRestrictionTest
Test: Manual verification with non-holdable sim to ensure that calls are
disconnected when receiving an incoming call on the other sim
Flag: com.android.server.telecom.flags.enable_call_sequencing

Change-Id: I6369cc2ca7a241cd5762bb3a1a3eb1bbabd52728