Set rejected transactional calls to disconnected before destroy
It was observed in a BR that if a call is rejected via a BT headset,
the ringer can get stuck and play indefinitely.

Upon investigation, this is result of not setting the call to
the disconnected state before removing the call from the BT ICS. This
caused the BT ICS to never update the headset.

Flag: EXEMPT high serverity bug with low regression chance
Bug: 397935575
Test: Unit + manual
Change-Id: Ica86b735e2c380d3f3c9538ce0bf1b9d2f687f5a