DSDA: Pending account selection multi-call aware
Ensure that the pending account selection is multi-call aware. Normally,
we clear the pending account selection future when we disconnect a call.
However, when we have a new call that is pending account selection and
we disconnect the active call, we will end up clearing the pending
account future for the new call and inadvertently cause that new call to
disconnect.

Bug: 396664319
Test: atest
CallsManagerTest#testPendingAccountSelectionNotClearedWithNewCall
Flag: EXEMPT bug fix

Change-Id: I1dafd06061ce072093d0f0b3b621d3ffc4a3d8f9