DSDA: Set simultaneous call type for metrics
Initial support for including metrics for call sequencing which will
track the simultaneous call type for each call. The types are based on a
priority which will be dynamically updated based on the newly added
calls. These will only be updated if the new simultaneous call type
priority is higher than the one already set for the call so that we can
have a better understanding on the metric side where sequencing may have
applied, if it did.

Bug: 384949208
Flag: com.android.server.telecom.flags.enable_call_sequencing
Test: atest CallsManagerTest#testSimultaneousCallType
Change-Id: I1251aac307f229d115e566e7161ef0051c1f22ae