Update preferred device strategy usage
Sync the preferred device strategy with the current communication device
if the corresponding values differ. Also ensure that we only use this
value at the start of a call. We have been observing many issues with
this API which are affecting the audio routing during calls and causing
audio to route incorrectly.

Bug: 377345692
Test: atest CallAudioRouteControllerTest
Test: Manual log verification
Flag: com.android.server.telecom.flags.update_preferred_audio_device_logic

Change-Id: I9986dccbc4b68830f156fff0b31adcc6a0d100dd