Fix flaky call audio route tests for auto cf
The tests on auto cf run mesasurably slower and as a result, we are
seeing an increase of flakiness on the call audio route tests. Several
tests are failing due to too many invocations (and inadvertently causes
subsequent tests to fail because the cleanup procedure isn't properly
enforced during failures). The other flakiness stems from not waiting
for the audio routing active state to change. We can leverage the new
CountDownLatches that were originally used for CUJ verification in the
unit tests to wait for the state change now.

Bug: 401278237
Test: atest CallAudioRouteControllerTest
Test: atest CallAudioRouteStateMachineTest
Flag: EXEMPT test fix
Change-Id: I81d92fd38c8b8ff87de0a7bd1e4420a7c8ab2ea8