Fix CallAudioRouteController tests multiple invocations
Another test fix for addressing flaky tests in CallAudioRouteController which arose as a result of auto cf running measurably slower. On verifications that have already been done before in the test, we can just add the atLeastOnce() operation just in case it gets counted twice in the verification.

Bug: 401534270
Test: atest CallAudioRouteControllerTest
Flag: EXEMPT test fix
Change-Id: I47b00525d4eafb67de7135077c1202322f2e5a83