DSDA: Modify max outgoing call restriction errors
Modifies how we portray the max outgoing call restrictions for when we
have an ongoing call being set up and we place another call. The message
should indicate that the user needs to wait for the call to be answered
or should disconnect it before placing a new call. Likewise, for when we
have a held + active call and try placing a 3rd outgoing call, we should
indicate to the user that merging the calls is not possible for
cross-sim based calling.

Bug: 403293836
Test: Manual test verification for both scenarios
Flag: EXEMPT bugfix
Change-Id: I7d7659fa576c2c8b69e0ea0085f4521b5edf7d35