DSDA: Add explicit handling for checking call type
Adds explicit handling for sequencing for determining if a call is
managed or a VOIP call (SM/transactional). Before, we were only checking
for self managed and also using that as a basis of determining if a
call is managed as well but we were disregarding checking that the call
was transactional.

Bug: 401051762
Test: atest CtsTelecomCujTestCases:CallSequencingMultiCallingTests
Test: atest CallSequencingTests
Flag: EXEMPT bugfix
Change-Id: I587eb5118980331ee8d0c73e4c7398538780d766