Use current active BT device in active routing
When we receive the BT_ACTIVE_DEVICE_PRESENT message in the
CallAudioRouteController, ensure that we only attempt to route into the
specified BT device if it's still the currently active device. When we
receive indication from the BT stack of what the new active device is
for any given BT profile, we update the active device cache stored in
CallAudioRouteController immediately before sending the message request
to process BT_ACTIVE_DEVICE_PRESENT. We have seen cases where by the
time the message is processed by the handler, the active device has
changed causing a potential loop when multiple BT devices are trying to
connect.

Bug: 389155187
Flag: EXEMPT bugfix
Test: atest CallAudioRouteControllerTest
Change-Id: I28004763cb21bc79f614f61ee41ec8d049d26d8d